Enable strict C++ compiler warnings and fix all violations
Add five new warning flags beyond the existing -Wall -Wextra -Werror:
- -Wshadow=local: catch dangerous local-to-local variable shadowing
- -Wnon-virtual-dtor: classes with virtual functions need virtual destructors
- -Wnull-dereference: potential null pointer dereferences
- -Wdouble-promotion: implicit float-to-double promotion
- -Wformat=2: format string security issues

Fix all resulting warnings across CUB, Thrust, libcudacxx, and cudax:
- Rename shadowing lambda parameters (args->inner_args, config->smem_config, etc.)
- Add explicit static_cast<double>() for intentional float-to-double conversions
- Add _CCCL_ASSERT for iterator validity to satisfy null-dereference checker
- Suppress false-positive null-dereference in basic_any internals via _CCCL_DIAG
- Suppress non-virtual-dtor from Catch2's ITransientExpression base class
- Suppress double-promotion in Thrust test assertions that compare mixed types
- Use float literals (0.00015f) when comparing against float variables

+++ cub/test/test_allocator.cu
@@ -444,9 +444,9 @@ int main(int argc, char** argv)
 
   printf("\t CUB CachingDeviceAllocator allocation CPU speedup: %.2f (avg cudaMalloc %.4f ms vs. avg DeviceAllocate "
          "%.4f ms)\n",
-         cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis,
-         cuda_malloc_elapsed_millis / timing_iterations,
-         cub_calloc_elapsed_millis / timing_iterations);
+         static_cast<double>(cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis),
+         static_cast<double>(cuda_malloc_elapsed_millis / timing_iterations),
+         static_cast<double>(cub_calloc_elapsed_millis / timing_iterations));
 
   // GPU performance comparisons.  Allocate and free a 1MB block 2000 times
   GpuTimer gpu_timer;
@@ -488,9 +488,9 @@ int main(int argc, char** argv)
 
   printf("\t CUB CachingDeviceAllocator allocation GPU speedup: %.2f (avg cudaMalloc %.4f ms vs. avg DeviceAllocate "
          "%.4f ms)\n",
-         cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis,
-         cuda_malloc_elapsed_millis / timing_iterations,
-         cub_calloc_elapsed_millis / timing_iterations);
+         static_cast<double>(cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis),
+         static_cast<double>(cuda_malloc_elapsed_millis / timing_iterations),
+         static_cast<double>(cub_calloc_elapsed_millis / timing_iterations));
 
   printf("Success\n");
